# services/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Customs API",
    "version": "0.1.0",
    "description": "Partial REST API contract for bootstrap and project token endpoints."
  },
  "components": {
    "schemas": {
      "BootstrapSetup": {
        "type": "object",
        "properties": {
          "firstTenantEnabled": {
            "type": "boolean"
          },
          "firstProxyEnabled": {
            "type": "boolean"
          },
          "defaultPoliciesEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "firstTenantEnabled",
          "firstProxyEnabled",
          "defaultPoliciesEnabled"
        ]
      },
      "PublicBootstrapStatus": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "state": {
            "type": "string",
            "enum": [
              "waiting_for_db",
              "schema_not_ready",
              "auth_unreachable",
              "no_users",
              "needs_setup",
              "ready"
            ]
          },
          "bundledMode": {
            "type": "boolean"
          },
          "setup": {
            "$ref": "#/components/schemas/BootstrapSetup"
          },
          "nextStep": {
            "type": "string",
            "enum": [
              "wait_for_runtime",
              "sign_in",
              "complete_setup",
              "done"
            ]
          },
          "ts": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "ok",
          "state",
          "bundledMode",
          "setup",
          "nextStep",
          "ts"
        ]
      },
      "DetailedBootstrapStatus": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PublicBootstrapStatus"
          },
          {
            "type": "object",
            "properties": {
              "checks": {
                "type": "object",
                "properties": {
                  "dbReady": {
                    "type": "boolean"
                  },
                  "schemaReady": {
                    "type": "boolean"
                  },
                  "authReachable": {
                    "type": "boolean"
                  },
                  "usersExist": {
                    "type": "boolean"
                  },
                  "ownerMembershipExists": {
                    "type": "boolean"
                  },
                  "tenantExists": {
                    "type": "boolean"
                  },
                  "placeholderTenantExists": {
                    "type": "boolean"
                  },
                  "bundledProxyConfigured": {
                    "type": "boolean"
                  },
                  "bundledProxyRegistered": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "dbReady",
                  "schemaReady",
                  "authReachable",
                  "usersExist",
                  "ownerMembershipExists",
                  "tenantExists",
                  "placeholderTenantExists",
                  "bundledProxyConfigured",
                  "bundledProxyRegistered"
                ]
              }
            },
            "required": [
              "checks"
            ]
          }
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "detail": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "code",
              "message",
              "detail"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "BootstrapFirstUserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "uuid"
              },
              "email": {
                "type": "string",
                "format": "email"
              }
            },
            "required": [
              "id",
              "email"
            ]
          }
        },
        "required": [
          "user"
        ]
      },
...
    },
...
  },
  "paths": {
    "/internal/bootstrap/status": {
      "get": {
        "tags": [
          "bootstrap"
        ],
        "summary": "Get public bootstrap status",
        "description": "Returns the coarse bootstrap state used by unauthenticated dashboard routing.",
        "responses": {
          "200": {
            "description": "Bootstrap status resolved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicBootstrapStatus"
                }
              }
            }
          },
          "503": {
            "description": "Bootstrap status resolved but the environment is degraded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicBootstrapStatus"
                }
              }
            }
          }
        }
      }
    },
    "/internal/bootstrap/status/detail": {
      "get": {
        "tags": [
          "bootstrap"
        ],
        "summary": "Get detailed bootstrap status",
        "description": "Returns the protected bootstrap diagnostics view for setup operations. Requires the x-bootstrap-secret header.",
        "responses": {
          "200": {
            "description": "Detailed bootstrap status resolved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedBootstrapStatus"
                }
              }
            }
          },
          "401": {
            "description": "Bootstrap secret is missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Detailed bootstrap status resolved but the environment is degraded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailedBootstrapStatus"
                }
              }
            }
          }
        }
      }
    },
    "/internal/bootstrap/first-user": {
      "post": {
        "tags": [
          "bootstrap"
        ],
        "summary": "Create the first bootstrap user",
        "description": "Creates the first dashboard user during environment setup. Requires the bootstrap secret header.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 128
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "First user created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapFirstUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bootstrap secret is missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "A first user already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Bootstrap is misconfigured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The auth admin backend returned an upstream error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Authentication service is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}